Improve data processing in Sql and Geofence models
Refactored some parts of the Sql and Geofence classes to handle data more efficiently and effectively. This involved introducing a temporary variable for storing converted value in Sql model, as well as optimizing the way temporary tables are created in the Geofence model. These improvements aim to enhance execution speed, code readability, and error realism.