Replace Model class usage with Job class in scheduler script
The code in the scheduler script has been refactored to use the Job class instead of the Model class. The function call to the external PHP script using numerous arguments was simplified to just calling a function within the Job class with the job id and database mame as parameters. This makes the code more maintainable by reducing complexity and minimizing the dependency on an external script.