The provided commit message does not correspond to the presented diffs. A more accurate commit message would be:
Add Location header and cleanup User.php

Refactored the code in User.php to add a "Location" header and changed the response structure for creating a user. The application now returns a Location header pointing to the new user's URL. It also improves the code clarity by returning response code as part of an array instead of an element of the user's data. The change in array syntax has also been cleaned up for readability and consistency.