Refactor function parameters and optimize regex in Model.php
This commit refactors the parameters of a function from strings to arrays in Model.php. It also optimizes a regular expression in this file to use lazy quantifiers, which can improve performance by reducing unnecessary backtracking. Basis for the optimization was switching from a greedy quantifier to a lazy one in a 'preg_match' function call.