Add handling for materialized foreign tables
A significant update was introduced that added several functionalities to handle materialized foreign tables. Firstly, the capacity to import a schema from a foreign server was added to the "Model" class allowing multiple schemas to be imported. Secondly, new endpoints were created to allow users to create, refresh, and drop materialized foreign tables. Lastly, conditional statements and forced parameters were introduced for better control when running a job in the scheduler.