Update MapServer repo and gc2-cli version in Dockerfile
In the Dockerfile, changed the repository for cloning MapServer from the main MapServer repository to the Mapcentia repository with branch 'rsvg_fix'. Also, updated the gc2-cli version to the 2024.3.3.79bafc6 version. The purpose is to keep the project up-to-date with the latest versions of software dependencies for optimal functionality.