Update Admin.php functions and add 'insertmeta' API endpoint
This commit includes the addition of a new public method 'get_insertmeta' to the 'Admin.php', extending the available API with an endpoint to handle default metadata for relations without any. Furthermore, return data types for several functions are simplified from 'array<mixed>' to 'array' for clarity, and the copyright year has been incremented.