Add health check in Dockerfile and remove unnecessary version argument
This commit includes a health check in Dockerfile of Mapserver_ecw5 which periodically checks if the application is running as expected. Additionally, it removes an unnecessary version argument in the Dockerfile of Mapserver that was previously used for installing gc2-cli. Removing unnecessary configurations ensures a cleaner, more maintainable codebase.