Update path to whitelist.txt in updatedb_whitelist.sh
Improved the updatedb_whitelist.sh script by using the SCRIPT_DIR variable to specify the absolute path of the whitelist.txt file. This change enhances the reliability of the script as it no longer depends on it being run from a specific directory. An error message is now also provided if whitelist.txt fails to open.