Improve cache clearing and settings retrieval
Refactored the clearCacheOnSchemaChanges method to clear cache using patterns instead of tags. Tweaked the get method of Setting.php to store cache per user because personal API keys are stored. Also, fixed a potential issue in Mapfile.php where a default extent is provided if the schema does not exist.