Update PostGIS image and adapt dependencies
The Docker files have been updated to use the mapcentia/postgis15 image instead of the previous mapcentia/postgis13. The PostGIS configuration and dependencies have been adapted to match this new version. Furthermore, the PHP version in the PostGIS Dockerfile is updated to PHP 8.3 and the Debian base image is changed to 'bookworm-slim' from 'bullseye-slim'.