Update from Postgres 13 to Postgres 15 and adjust health checks
This commit upgrades Postgres from version 13 to 15 in the docker-compose.yml file, the Dockerfile and the entrypoint.sh script. In addition, the timeout and start period for health checks in Dockerfile have been reduced to 10 seconds. Changes were also made to the pg_hba.conf and supervisord.conf files to reflect the Postgres update.