Handle PDOException in User API
Added a try/catch block in the User API to handle PDOException that may arise when running the createSchema method. Also updated the createSchema method's docblock in the Database model to reflect that it throws PDOException. This change will prevent the application from completely stopping when an exception is thrown during database schema creation.