Updated field assignment and added cache deletion in get.php
In the get.php file, the field assignment method for 'fieldType' and 'fieldValue' was updated by taking advantage of null coalescing operator, to make the code cleaner and more efficient. Furthermore, an action to clear the cache using specific patterns based on the current table, schema, and name was also added.