Reintroduce v1 SQL API with updated logic
In context of the refactoring, the previously removed v1 SQL API has been reintroduced with the updated logic. Modifications have been implemented in the routing method in index.php file to accommodate the changes. Additionally, a new class dedicated for SQL operations under v1 of the API is added to handle/index and post_index requests.